Fix s64 and u64 defines

+++ src/utils/defines.h
@@ -16,17 +16,17 @@
 #define DESHI_NAME_SIZE 64
 
 //number typedefs
-typedef signed char    s8;
-typedef signed short   s16;
-typedef signed int     s32;
-typedef signed long    s64;
-typedef unsigned char  u8;
-typedef unsigned short u16;
-typedef unsigned int   u32;
-typedef unsigned long  u64;
-typedef float          f32;
-typedef double         f64;
-typedef s32            b32; //int-based boolean so c++ doesnt convert to 0 or 1
+typedef signed char        s8;
+typedef signed short       s16;
+typedef signed int         s32;
+typedef signed long long   s64;
+typedef unsigned char      u8;
+typedef unsigned short     u16;
+typedef unsigned int       u32;
+typedef unsigned long long u64;
+typedef float              f32;
+typedef double             f64;
+typedef s32                b32; //int-based boolean so c++ doesnt convert to 0 or 1
 
 //static defines
 #define static_internal static
